requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
python-dotenv==0.12.0
python-telegram-bot==12.4.2
requests==2.23.0
terminaltables==3.1.0
pandas==1.0.2
plotly==4.5.4
